Selective trigger un-publish for C/U/D
Includes clean up of published MQTT read/write topics.

Selective un-publish is for triggers with object keyword "lvl", where levels are removed from the list. This selective un-publish is not yet implemented on script re-start (for where checkForChanges = false), but that is on the to-do list.

Publishing 'junk' to topics that may not have been published yet may have unforeseen consequences, so this is experimental at this stage.  Some value / junk is needed to ensure the topic does actually get deleted, either that or the currently published values need to be tracked, which is not done by this script for cbus/write topics.

Selective trigger un-publish for C/U/D

Includes clean up of published MQTT read/write topics.

Selective un-publish is for triggers with object keyword "lvl", where levels are removed from the list. This selective un-publish is not yet implemented on script re-start (for where checkForChanges = false), but that is on the to-do list.

Publishing 'junk' to topics that may not have been published yet may have unforeseen consequences, so this is experimental at this stage. Some value / junk is needed to ensure the topic does actually get deleted, either that or the currently published values need to be tracked, which is not done by this script for cbus/write topics.