Ditch 'fanKey' and "fix" sweep fan control
Sweep fans always require an explicit level to be stated. "Hey Google set the fan speed to low".

To use a 'lastlevel', a setting of high speed (aka 255 / ON) would not be set if the lastlevel logic were used because it would be ignored and the last level set.

This is not really what I want, as high speed is rarely used but might be, so I would want the fan to return to its last level usually, but I'm unable to find a better solution than excluding high speed as an option.

On the plus side, the variable 'fanKey' is history.