Rewrite approach to ramping #8
Instructs the bridge to use a V1 API 'transitiontime'. A known issue is that when ramped off using this method it will cause altering the last level that a light was set to. A subsequent "turn on" will set the brightness to the last status update received during ramp to off, which may be a very low value.